Submitted code for progress bar, asteroid spawning and fixes for unlock / lock nodes
Submitted code to spawn asteroids at the start of the level based on the
size and color distributions in the levels file.

Submitted code for progress bar which increments every time an asteroid
is destroyed. Increment amount is based on total rocks for the level.

Also made fixes to lock unlock animation.